Ticket 48805 - Sign comparison checks.
Bug Description:  We perform a lot of checks on values that do not have the
correct or same sign type. This often indicates the incorrect type of the
variable, or incorrect use and design of the application.

Fix Description:
* Cast some variables as needed where the type cannot be changed.
* Change the type of some variables where correct.
* Ignore others due to fundamental library issues.

https://fedorahosted.org/389/ticket/48805

Author: wibrown

Review by: mreynolds, nhosoi (Thanks!)

+++ ldap/servers/plugins/rootdn_access/rootdn_access.c
@@ -691,7 +691,7 @@ rootdn_check_host_wildcard(char *host, char *client_host)
 static int
 rootdn_check_ip_wildcard(char *ip, char *client_ip)
 {
-    int ip_len = strlen(ip);
+    size_t ip_len = strlen(ip);
     int i;
     /*
      *  Start at the beginning of the string and move forward, and skip the last char "*"
@@ -711,7 +711,7 @@ rootdn_check_ip_wildcard(char *ip, char *client_ip)
 
 char *
 strToLower(char *str){
-    int i;
+    size_t i;
 
     for(i = 0; str && i < strlen(str); i++){
         str[i] = tolower(str[i]);
